+++ lib/msal-node/src/client/PublicClientApplication.ts
@@ -55,10 +55,7 @@ export class PublicClientApplication extends ClientApplication implements IPubli
      */
     public async acquireTokenByDeviceCode(request: DeviceCodeRequest): Promise<AuthenticationResult | null> {
         this.logger.info("acquireTokenByDeviceCode called");
-        const validRequest: CommonDeviceCodeRequest = {
-            ...request,
-            ...this.initializeBaseRequest(request)
-        };
+        const validRequest: CommonDeviceCodeRequest = Object.assign(request, this.initializeBaseRequest(request));
         const serverTelemetryManager = this.initializeServerTelemetryManager(ApiId.acquireTokenByDeviceCode, validRequest.correlationId!);
         try {
             const deviceCodeConfig = await this.buildOauthClientConfiguration(