+++ lib/msal-angular/changelog.md
@@ -1,8 +1,35 @@
 # Change Log - @azure/msal-angular
 
-This log was last generated on Mon, 15 Mar 2021 23:45:17 GMT and should not be manually modified.
+This log was last generated on Wed, 14 Apr 2021 18:39:53 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 2.0.0-beta.4
+
+Wed, 14 Apr 2021 18:39:53 GMT
+
+### Changes
+
+- fix(Msal-Angular): Dynamic interceptor authority (1292510+svrooij@users.noreply.github.com)
+
+## 2.0.0-beta.3
+
+Wed, 31 Mar 2021 22:25:57 GMT
+
+### Changes
+
+- Update MsalInterceptor types and getScopesForEndpoint with query and relative paths #3307 (joarroyo@microsoft.com)
+
+## 2.0.0-beta.2
+
+Wed, 24 Mar 2021 22:55:46 GMT
+
+### Changes
+
+- Add support for logoutPopup API #3044 (thomas.norling@microsoft.com)
+- Add MsalCustomNavigationClient, updates samples and docs #3174 (joarroyo@microsoft.com)
+- Update MsalGuard and _inProgress subject #3269 (joarroyo@microsoft.com)
+- Update MSAL Service handleRedirectObservable hash handling #3243 (joarroyo@microsoft.com)
 
 ## 2.0.0-beta.1